% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellBarcodeSim.R
\docType{package}
\name{CellBarcodeSim}
\alias{CellBarcodeSim}
\title{CellBarcodeSim}
\description{
This package aim to simulate DNA barcode sequencing results,
including both UMI (unique molecular identifier) sequencing and non-UMI
sequencing techniques.
}
