% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_simulation.R
\name{sample_barcode}
\alias{sample_barcode}
\title{Function sample barcode library}
\usage{
sample_barcode(n, d_barcode_lib = NULL)
}
\arguments{
\item{n}{number of barcodes to sample}

\item{d_barcode_lib}{a data.frame with two columns: \code{seq} and \code{freq}.}
}
\value{
a data frame
}
\description{
Function sample barcode library
}
