% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_simulation.R
\name{simu_barcode_hamming_exp}
\alias{simu_barcode_hamming_exp}
\title{Simulate hamming distance barcode library with exponential distribution}
\usage{
simu_barcode_hamming_exp(
  length = 10,
  dist = 3,
  n = 1e+06,
  rate = 1,
  output = "hamming_barcodes_exp.tsv"
)
}
\description{
Simulate hamming distance barcode library with exponential distribution
}
