% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_simulation.R
\name{simu_barcode_hamming_uniform}
\alias{simu_barcode_hamming_uniform}
\title{Simulate hamming distance barcode library with uniform distribution}
\usage{
simu_barcode_hamming_uniform(
  length = 10,
  dist = 3,
  output = "hamming_barcodes.tsv"
)
}
\description{
Simulate hamming distance barcode library with uniform distribution
}
