% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_simulation.R
\name{simu_barcode_random_exp}
\alias{simu_barcode_random_exp}
\title{Simulate random barcode library with exponential distribution}
\usage{
simu_barcode_random_exp(
  length = 14,
  n = 1e+06,
  rate = 1,
  output = "random_barcodes_exp.tsv"
)
}
\description{
Simulate random barcode library with exponential distribution
}
