% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_simulation.R
\name{simu_barcode_random_lnorm}
\alias{simu_barcode_random_lnorm}
\title{Simulate random barcode library with log normal distribution}
\usage{
simu_barcode_random_lnorm(
  length = 14,
  n = 1e+06,
  log_mean = 1,
  log_sd = 1,
  output = "random_barcodes_lnorm.tsv"
)
}
\description{
Simulate random barcode library with log normal distribution
}
