% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_simulation.R
\name{simu_barcode_random_norm}
\alias{simu_barcode_random_norm}
\title{Simulate random barcode library with normal distribution}
\usage{
simu_barcode_random_norm(
  length = 14,
  n = 1e+06,
  mean = 1,
  sd = 1,
  output = "random_barcodes_norm.tsv"
)
}
\description{
Simulate random barcode library with normal distribution
}
