% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_simulation.R
\name{simu_barcode_random_uniform}
\alias{simu_barcode_random_uniform}
\title{Simulate random barcode library with uniform distribution}
\usage{
simu_barcode_random_uniform(
  length = 14,
  n = 1e+06,
  top_fix = "",
  bottom_fix = "",
  output = "random_barcodes_uniform.tsv"
)
}
\arguments{
\item{length}{barcode length with default 14}

\item{n}{number of barcodes with default 1e6}

\item{top_fix}{5 end fix region sequence of the barcode with default "".}

\item{bottom_fix}{3 end fix region sequence of the barcode with default "".}

\item{output}{output file name with default "random_barcodes_uniform.tsv"}
}
\value{
a barcode library file or a data frame
}
\description{
Simulate random barcode library with uniform distribution
}
