% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_simulation.R
\name{simu_clone_size_lognormal}
\alias{simu_clone_size_lognormal}
\title{Simulate clone size with log normal distribution}
\usage{
simu_clone_size_lognormal(n, size_mean = 1.2, size_variant = 2)
}
\arguments{
\item{n}{clone number}

\item{size_mean}{meanlog of log normal distribution with default 1.2}

\item{size_variant}{meansd of log normal distribution with default 2}
}
\value{
a vector of clone size
}
\description{
Simulate clone size with log normal distribution
}
