% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_simulation.R
\name{simu_clone_size_powerlaw}
\alias{simu_clone_size_powerlaw}
\title{Simulate clone size with power law distribution}
\usage{
simu_clone_size_powerlaw(n, constant = 10, scale = 1, alpha = 2)
}
\arguments{
\item{n}{clone number}

\item{constant}{constant of power law distribution with default 10}

\item{scale}{scale of power law distribution with default 1}

\item{alpha}{alpha of power law distribution with default 2}
}
\value{
a vector of clone size
}
\description{
Simulate clone size with power law distribution
}
