% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_simulation.R
\name{simu_clone_size_uniform}
\alias{simu_clone_size_uniform}
\title{uniform:
parameters:
max clone size
min clone size
clone number}
\usage{
simu_clone_size_uniform(n, size_max = 1000, size_min = 1)
}
\description{
uniform:
parameters:
max clone size
min clone size
clone number
}
