% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_simulation.R
\name{simu_sequence_run_command}
\alias{simu_sequence_run_command}
\title{Need to download the ART to the \code{lib/ngs_simu} fold.
MiSeq sequencing model
parameters
input: fasta file
profile: it can be MSv1, HS20 ...
reads_length: default is 110
output: the output of the simulated sequencing reads}
\usage{
simu_sequence_run_command(
  input,
  profile,
  reads_length = 110,
  output,
  art_bin = NULL
)
}
\description{
Need to download the ART to the \code{lib/ngs_simu} fold.
MiSeq sequencing model
parameters
input: fasta file
profile: it can be MSv1, HS20 ...
reads_length: default is 110
output: the output of the simulated sequencing reads
}
