% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_simulation.R
\name{simulate_main}
\alias{simulate_main}
\title{Run the simulation}
\usage{
simulate_main(
  barcode_library_file = NULL,
  clone_size_dist = "uniform",
  clone_n = 300,
  clone_size_dist_par = list(size_max = 1000, size_min = 1),
  cycle = 30,
  efficiency = 0.705,
  error = 1e-06,
  pcr_read_per_cell = 50,
  output_prefix = "./tmp/simu_seq",
  ngs_profile = "MSv1",
  reads_length = 100,
  is_replicate = F,
  top_seq = "",
  bottom_seq = "",
  sequence_trunk = 10,
  art_bin = NULL
)
}
\description{
Run the simulation
}
