% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_simulation.R
\name{simulate_main_umi}
\alias{simulate_main_umi}
\title{Run the simulation for UMI_barcode sequencing}
\usage{
simulate_main_umi(
  barcode_library_file = NULL,
  barcode_library = NULL,
  clone_size_dist = "uniform",
  clone_n = 300,
  clone_size_dist_par = list(size_max = 1000, size_min = 1),
  cycle = 40,
  efficiency = 0.705,
  error = 1e-06,
  pcr_read_per_umi = 50,
  output_prefix = "./tmp/simu_umi_seq",
  ngs_profile = "MSv1",
  reads_length = 100,
  top_seq = "",
  bottom_seq = "",
  sequence_trunk = 10,
  preamp_n = 0,
  umi_length = 8,
  umi_tagging_efficiency = 0.25,
  qc_shift = 0,
  art_bin = NULL
)
}
\arguments{
\item{barcode_library_file}{barcode library file, if not provided, the barcode_library should be provided.}

\item{barcode_library}{barcode library data frame with two columns: \code{seq} and \code{freq}.}

\item{clone_size_dist}{clone size distribution, should be \code{uniform}, \code{lognormal} or \code{powerlaw}.}

\item{clone_n}{number of clones.}

\item{clone_size_dist_par}{parameters for clone size distribution, check the function \code{\link{simu_clone_size_uniform}}, \code{\link{simu_clone_size_lognormal}}, and \code{\link{simu_clone_size_powerlaw}}.}

\item{cycle}{total PCR cycle count with default 40.}

\item{efficiency}{pcr efficiency, default is 0.705.}

\item{error}{pcr error per base per cycle with default 1e-6.}

\item{pcr_read_per_umi}{reads number sampled from PCR results for sequencing with default 50.}

\item{output_prefix}{output prefix of the simulated sequencing result with default "./tmp/simu_umi_seq".}

\item{ngs_profile}{sequencing profile with default "MSv1".}

\item{reads_length}{reads length of the simulated sequencing result with default 100.}

\item{top_seq}{5 end fix region sequence added after PCR default "".}

\item{bottom_seq}{3 end fix region sequence added after PCR default "".}

\item{sequence_trunk}{the max length of the barcode sequence to be used for the simulation with default 10.}

\item{preamp_n}{pre-amplification cycle number with default 0.}

\item{umi_length}{UMI length with default 8.}

\item{umi_tagging_efficiency}{UMI tagging efficiency with default 0.25.}

\item{qc_shift}{quality score shift, default is 0.}

\item{art_bin}{path to the ART bin directory, default is NULL, which will use the package buildin ART bin directory.}
}
\value{
a list with two items: \code{library_fasta} is the fasta file of the simulated library; \code{sequencing_result} is the prefix of the simulated sequencing result.
}
\description{
Run the simulation for UMI_barcode sequencing
}
