% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_simulation.R
\name{simu_barcode_hamming_lnorm}
\alias{simu_barcode_hamming_lnorm}
\title{Simulate hamming distance barcode library with log normal distribution}
\usage{
simu_barcode_hamming_lnorm(
  length = 10,
  dist = 3,
  n = 1e+06,
  log_mean = 1,
  log_sd = 1,
  output = "hamming_barcodes_lnorm.tsv"
)
}
\description{
Simulate hamming distance barcode library with log normal distribution
}
