% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_simulation.R
\name{simu_barcode_hamming_norm}
\alias{simu_barcode_hamming_norm}
\title{Simulate hamming distance barcode library with normal distribution}
\usage{
simu_barcode_hamming_norm(
  length = 10,
  dist = 3,
  n = 1e+06,
  mean = 1,
  sd = 1,
  output = "hamming_barcodes_norm.tsv"
)
}
\description{
Simulate hamming distance barcode library with normal distribution
}
