% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_simulation.R
\name{simu_barcode_random_lnorm}
\alias{simu_barcode_random_lnorm}
\title{Simulate random barcode library with log normal distribution}
\usage{
simu_barcode_random_lnorm(
  length = 14,
  n = 1e+06,
  log_mean = 1,
  log_sd = 1,
  top_fix = "",
  bottom_fix = "",
  output = "random_barcodes_lnorm.tsv"
)
}
\arguments{
\item{length}{barcode length with default 14}

\item{n}{number of barcodes with default 1e6}

\item{log_mean}{meanlog of log normal distribution with default 1}

\item{log_sd}{sdlog of log normal distribution with default 1}

\item{top_fix}{5 end fix region sequence of the barcode with default "".}

\item{bottom_fix}{3 end fix region sequence of the barcode with default "".}

\item{output}{output file name with default "random_barcodes_lnorm.tsv"}
}
\value{
a barcode library file or a data frame
}
\description{
Simulate random barcode library with log normal distribution
}
