% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_simulation.R
\name{simu_barcode_random_uniform}
\alias{simu_barcode_random_uniform}
\title{Simulate random barcode library with uniform distribution}
\usage{
simu_barcode_random_uniform(
  length = 14,
  n = 1e+06,
  output = "random_barcodes_uniform.tsv"
)
}
\description{
Simulate random barcode library with uniform distribution
}
