% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_simulation.R
\name{simu_clone_size_lognormal}
\alias{simu_clone_size_lognormal}
\title{log normal:
parameters:
clone size variation
clone number
average clone size}
\usage{
simu_clone_size_lognormal(n, size_mean = 1.2, size_variant = 2)
}
\description{
log normal:
parameters:
clone size variation
clone number
average clone size
}
